Add repositories for AuditLog, WaterBill, BillPayment, InsuranceBill, ElectricityBill, and TelecommunicationBill; refactor UserAccountDTO and Loan models to use int for userId and id

+++ src/main/java/com/customer_backend/models/Loan.java
@@ -13,10 +13,10 @@ public class Loan {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "user_id")
-    private Long userId;
+    private int userId;
 
     @Column(name = "loan_amount")
     private Double loanAmount;
@@ -31,19 +31,19 @@ public class Loan {
     private String status;
 
     // Getters and Setters
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
